Refactor: Moved backfil to admin commands
